modules/tools/automated-tests/requirements.txt
cognite-sdk>=6.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-html>=4.0.0
